database: api to add game entries

+++ src/shiratsu-lib/src/parse/generic/dat.rs
@@ -9,7 +9,7 @@ use super::super::*;
 #[derive(Debug, Deserialize, PartialEq)]
 struct Rom {
     name: String,
-    size: u64,
+    size: u32,
     crc: String,
     md5: String,
     sha1: String,
